// src/types/member.ts
export interface Member {
    memberName: string;
    status: string;
    userCreate: string;
    dateUpd: string;
    userUpd: string;
    vehicleType: string;
    vehicleNo: string;
    dateCreate: string;
    memberId: string;
    startDate: string;
    vehicleUnit: string;
    expiredDate: string;
    vehicleDesc: string;
}